Added a player with a walking animation, fixed scrolling system
The scrolling system is now very fluid - the player takes roughly half a
second to fully accelerate to full speed, change direction or stop.
Added a player sprite, with a basic spritesheet, that animates walking.
Added an fps debug, and made other debugs easier to read. Fixed the
title of the page to 'Aiopa RPG'.